### 2019-01-29 #43 @whedon
[PRE REVIEW]: Aero Python: classical aerodynamics of potential flow using Python
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/AeroPython" target ="_blank">https://github.com/barbagroup/AeroPython</a>
**Version:** 0.0
