+++ paper.bib
@@ -1,8 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-
-%% Created for Lorena A Barba at 2019-01-28 21:02:33 -0500 
+%% Created for Lorena Barba at 2019-01-29 12:56:30 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -12,9 +11,10 @@
 @misc{Barba2017,
 	Author = {Barba, Lorena A.,},
 	Date-Added = {2019-01-28 20:57:11 -0500},
-	Date-Modified = {2019-01-28 20:58:15 -0500},
+	Date-Modified = {2019-01-29 12:56:30 -0500},
 	Howpublished = {figshare, \href{https://doi.org/10.6084/m9.figshare.4584328.v1}{doi:10.6084/m9.figshare.4584328.v1}},
-	Title = {MAE-6226: Aerodynamics Course Syllabus}}
+	Title = {MAE-6226: Aerodynamics Course Syllabus},
+	Year = {2017}}
 
 @article{margulieuxETal2016,
 	Author = {Margulieux, Lauren E and Catrambone, Richard and Guzdial, Mark},
@@ -81,7 +81,8 @@
 	Series = {Lecture Notes in Engineering},
 	Title = {{XFOIL}: An analysis and design system for low {Reynolds} number airfoils},
 	Volume = {54},
-	Year = {1989}}
+	Year = {1989},
+	Bdsk-Url-1 = {https://doi.org/10.1007/978-3-642-84010-4_1}}
 
 @article{BarbaForsyth2018,
 	Author = {Barba, Lorena A., and Forsyth, Gilbert F.},
