### 2019-01-29 #43 @labarba
👋 @kyleniemeyer — This is the second JOSE submission from Barbagroup. Thanks for editing!
