### 2019-01-29 #43 @labarba
As for reviewers, @Juanlu001 just finished a review for JOSS, but he'd be great for this one; I hope he'd be willing to volunteer the time 😁
If not, maybe he can suggest someone else from the @AeroPython network? (name coincidence! see [tweet](https://twitter.com/LorenaABarba/status/464041427169583104) from 2014)
