### 2019-01-30 #43 @weymouth
I'm happy to do a review

Gabriel D Weymouth
_______________________________________________
"Computers are useless. They can only give you answers." Pablo Picasso

On Tue, Jan 29, 2019 at 11:30 PM Lorena A. Barba <notifications@github.com>
wrote:

> Another member of the @AeroPython <https://github.com/AeroPython> group
> who could be interested in reviewing this is @AlexS12
> <https://github.com/AlexS12>.
>
> Browsing the forks, I found that @weymouth <https://github.com/weymouth>
> has a course on Marine Hydrodynamics
> <https://github.com/weymouth/MarineHydro> that took some inspiration from
> our AeroPython module. He could be willing to contribute a review.
>
> Also, a Russian researcher and CFD specialist at TsAGI, Innokentiy
> Kursakov, is working on a translation to Russian
> <https://github.com/ikursakov/ru_AeroPython>—he may be willing to be an
> official reviewer!
>
