### 2019-01-31 #45 @whedon
[REVIEW]: Aero Python: classical aerodynamics of potential flow using Python
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/AeroPython" target ="_blank">https://github.com/barbagroup/AeroPython</a>
**Version:** v1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.1004727.v4" target="_blank">10.6084/m9.figshare.1004727.v4</a>
