### 2019-02-28 30cd852
Consolidating for 2019
Attempting to simplify this down to 4 lectures.
