+++ lessons/BoundaryLayer.py
@@ -1,17 +1,14 @@
-""" Solve boundary layer problems using Pohlhausen's method
-
-This module holds routines to determine the evolution of a boundary layer
-in a varying external flow assuming a Pohlhausen velocity profile.
+""" Solve boundary layer problems 
 
 Methods:
     disp_ratio, mom_ratio, df_0
-    march, sep
+    thwaites, sep
 
-Imports: numpy, cumtrapz & odeint from scipy.integrate
+Imports: numpy, cumtrapz from scipy.integrate, ceil from math
 """
 
-import numpy
-from scipy.integrate import cumtrapz,odeint
+import numpy as np
+from scipy.integrate import cumtrapz
 
 # displacement thickness ratios
 def disp_ratio(lam): return 3./10.-lam/120.
@@ -23,12 +20,11 @@ def mom_ratio(lam): return 37./315.-lam/945.-lam**2/9072.
 def df_0(lam): return 2+lam/6.
 
 # look-up table for lambda = f(lambda_2)
-_lam_range = numpy.linspace(-17.5,12)
+_lam_range = np.linspace(-17.5,12)
 _lam2_range = _lam_range*mom_ratio(_lam_range)**2
-def _lam2_inv(lam2): return numpy.interp(lam2,_lam2_range,_lam_range)
 
-def march(s,u_s,nu,thwaites=False):
-    """ March the boundary layer momentum integral equation
+def thwaites(s,u_s):
+    """ Integrate Thwaites over the boundary layer and find shape factor
 
     Notes:
     Output array values after separation(lam<-12) are meaningless.
@@ -36,55 +32,38 @@ def march(s,u_s,nu,thwaites=False):
     Inputs:
     s   -- array of distances along the boundary layer; must be positive and increasing
     u_s -- array of external velocities; must be positive
-    nu  -- kinematic viscosity; must be scalar
-    thwaites -- Thwaites approximate method flag (default=False)
 
     Outputs:
     delta2 -- momentum thickness array
-    lam   -- shape parameter array
-    iSep  -- separation index
+    lam    -- shape factor array
+    iSep   -- distance along array to separation point (iSep=len(s) means no separation)
 
     Examples:
-    s = numpy.linspace(0,numpy.pi,16)        # define distance array
-    u_s = 2*numpy.sin(s)                     # define external velocity (circle example)
-    delta2,lam,iSep = march(s,u_s,nu=1e-5)   # march along to the point of separation
+    s = np.linspace(0,np.pi,16)      # define distance array
+    u_s = 2*np.sin(s)                # define external velocity (circle example)
+    delta2,lam,iSep = march(s,u_s)   # march along to the point of separation
     """
-    # velocity gradient
-    du_s = numpy.gradient(u_s)/numpy.gradient(s)
+    # Thwaites approximation for delta_2^2
+    delta22 = 0.45/u_s**6*cumtrapz(u_s**5,s,initial=0)
 
-    if thwaites:
-        # Thwaites approximation for delta_2^2
-        delta22 = 0.45*nu/u_s**6*cumtrapz(u_s**5,s,initial=0)
+    # adjust for (Blasius) flat plate IC
+    if du_s[0]==0: delta22 += 0.441*nu/u_s[0]*s[0]
 
-        # adjust for (Blasius) flat plate IC
-        if du_s[0]==0: delta22 += 0.441*nu/u_s[0]*s[0]
-
-    else:
-        # define Polhausen momentum integral ODE
-        def func(y,t): # delta_2^2 => y, s => t
[…]
