+++ lessons/VortexPanel.py
@@ -9,15 +9,15 @@ Classes:
 
 Methods:
     panelize, concatenate
-    make_polygon, make_ellipse, make_circle, make_jukowski
+    make_ellipse, make_circle, make_jfoil
 
 Imports:
     numpy, pyplot from matplotlib, march & sep from BoundaryLayer
 """
 
-import numpy
-from matplotlib import pyplot
-from BoundaryLayer import march,sep
+import numpy as np
+from matplotlib import pyplot as plt
+from BoundaryLayer import thwaites,sep
 
 ### Fundamentals
 
@@ -50,7 +50,7 @@ class Panel(object):
         self.gamma = gamma; self._gamma = (gamma,gamma) # copy gamma
         self.xc = 0.5*(x0+x1); self.yc = 0.5*(y0+y1)    # panel center
         dx = x1-self.xc; dy = y1-self.yc
-        self.S = numpy.sqrt(dx**2+dy**2)                # half-width
+        self.S = np.sqrt(dx**2+dy**2)                # half-width
         self.sx = dx/self.S; self.sy = dy/self.S        # tangent
 
     def velocity(self, x, y):
@@ -84,17 +84,17 @@ class Panel(object):
         my_panel = vp.Panel(0,-1,0,1)  # creates panel on y-axis
         my_panel.plot()                # plot the panel
         """
-        return pyplot.plot(self.x, self.y, style, lw=2)
+        return plt.plot(self.x, self.y, style, lw=2)
 
     def _constant(self, x, y):
         "Constant panel induced velocity"
         lr, dt, _, _ = self._transform_xy(x, y)
-        return self._rotate_uv(-dt*0.5/numpy.pi, -lr*0.5/numpy.pi)
+        return self._rotate_uv(-dt*0.5/np.pi, -lr*0.5/np.pi)
 
     def _linear(self, x, y):
         "Linear panel induced velocity"
         lr, dt, xp, yp = self._transform_xy(x, y)
-        g, h, c = (yp*lr+xp*dt)/self.S, (xp*lr-yp*dt)/self.S+2, 0.25/numpy.pi
+        g, h, c = (yp*lr+xp*dt)/self.S, (xp*lr-yp*dt)/self.S+2, 0.25/np.pi
         return (self._rotate_uv(c*( g-dt), c*( h-lr))
                +self._rotate_uv(c*(-g-dt), c*(-h-lr)))
 
@@ -103,8 +103,8 @@ class Panel(object):
         xt = x-self.xc; yt = y-self.yc # shift x,y
         xp = xt*self.sx+yt*self.sy     # rotate x
         yp = yt*self.sx-xt*self.sy     # rotate y
-        lr = 0.5*numpy.log(((xp-self.S)**2+yp**2)/((xp+self.S)**2+yp**2))
-        dt = numpy.arctan2(yp,xp-self.S)-numpy.arctan2(yp,xp+self.S)
+        lr = 0.5*np.log(((xp-self.S)**2+yp**2)/((xp+self.S)**2+yp**2))
+        dt = np.arctan2(yp,xp-self.S)-np.arctan2(yp,xp+self.S)
         return lr, dt, xp, yp
 
     def _rotate_uv(self, up, vp):
@@ -118,7 +118,7 @@ class PanelArray(object):
     """Array of vortex panels
 
     Attributes:
-    panels -- the numpy array of panels
+    panels -- the np array of panels
     alpha  -- the flow angle of attack
     """
 
@@ -126,7 +126,7 @@ class PanelArray(object):
         """Initialize a PanelArray
 
         Inputs:
-        panels -- a numpy array of panels
+        panels -- a np array of panels
 
         Outputs:
         A PanelArray object.
@@ -156,7 +156,7 @@ class PanelArray(object):
[…]
