### 2019-03-01 8470f70
Cleaned up through second tutorial
