### 2019-03-04 41af5f0
Update VortexPanel.py
