+++ lessons/VortexPanel.py
@@ -220,7 +220,7 @@ class PanelArray(object):
         # construct the matrix
         A = np.zeros((len(xc), len(xc)))         # empty matrix
         for j, p_j in enumerate(self.panels):    # loop over panels
-            u0,v0,u1,v1 = p_j._linear(xc,yc)        # f_j at all panel centers
+            u0,v0,u1,v1 = p_j.linear(xc,yc)        # f_j at all panel centers
             A[:,self.left[j]] += -u0*sy+v0*sx       # -S end influence
             A[:,j] += -u1*sy+v1*sx                  # +S end influence
 
