### 2019-03-07 17ed6c2
Improved cumtrap for powers
