+++ README.md
@@ -20,12 +20,10 @@ Teaching numerical hydrodynamics with [Jupyter notebooks](http://jupyter.org/) w
 * [Quick Python Intro](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/0_0_QuickPythonIntro.ipynb)
 
 #### Numerical methods
-1. [Vortex panels](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_1_VortexPanel.ipynb)
-1. [Solution method](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_2_SolutionMethod.ipynb)
-1. [Lifting bodies](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_3_LiftingBodies.ipynb)
-1. [Boundary layer solver](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_4_BoundaryLayerSolver.ipynb)
-1. [Separation prediction](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_5_SeparationPrediction.ipynb)
-1. [Linear panels and pressure Drag](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_6_PressureDrag.ipynb)
+1. [Vortex panel solution method](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_1_SolutionMethod.ipynb)
+1. [Lifting bodies and linear panels](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_2_LiftingBodies.ipynb)
+1. [Boundary layers and separation prediction](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_3_SeparationPrediction.ipynb)
+1. [Pressure Drag](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/lessons/3_4_PressureDrag.ipynb)
 
 #### Assignment
 * [Overview](http://nbviewer.ipython.org/urls/github.com/weymouth/MarineHydro/blob/master/coursework3/Assignment.ipynb)
