### 2019-03-12 ef6f951
Update VortexPanel.py
