+++ lessons/VortexPanel.py
@@ -9,7 +9,7 @@ Classes:
 
 Methods:
     panelize, concatenate
-    make_ellipse, make_circle, make_jfoil
+    make_ellipse, make_circle, make_jfoil, make_spline
 
 Imports:
     numpy, pyplot from matplotlib, march & sep from BoundaryLayer
@@ -42,7 +42,7 @@ class Panel(object):
         Outputs:
         A Panel object.
 
-        Examples:
+        Example:
         p_1 = vp.Panel(-1,0,1,0)    # make panel on x-axis with gamma=0
         p_2 = vp.Panel(0,-1,0,1,4)  # make panel on y-axis with gamma=4
         """
@@ -50,8 +50,12 @@ class Panel(object):
         self.gamma = gamma; self._gamma = (gamma,gamma) # copy gamma
         self.xc = 0.5*(x0+x1); self.yc = 0.5*(y0+y1)    # panel center
         dx = x1-self.xc; dy = y1-self.yc
-        self.S = np.sqrt(dx**2+dy**2)                # half-width
+        self.S = np.sqrt(dx**2+dy**2)                   # half-width
         self.sx = dx/self.S; self.sy = dy/self.S        # tangent
+        if np.isclose(self.S,0):                        # check panel size
+            raise ValueError("Panels must have non-zero length.\n"+
+                             "Your endpoints are on top of each other.")
+
 
     def velocity(self, x, y):
         """Compute the velocity induced by the panel
@@ -62,7 +66,7 @@ class Panel(object):
         Outputs:
         u,v   -- the x and y components of the velocity
 
-        Examples:
+        Example:
         p_2 = vp.Panel(0,-1,0,1,4)        # make panel on y-axis with gamma=4
         u,v = p_2.velocity(-1,0)          # get induced velocity on x-axis
         """
@@ -80,7 +84,7 @@ class Panel(object):
         Inputs:
         style -- a string defining the matplotlib style
 
-        Examples:
+        Example:
         my_panel = vp.Panel(0,-1,0,1)  # creates panel on y-axis
         my_panel.plot()                # plot the panel
         """
@@ -137,6 +141,13 @@ class PanelArray(object):
         self.bodies = [(0,n)]        # range for a body
         self.left = [n-1]+list(range(n-1)) # index to the left
 
+        # compute area
+        yc, S, sx = self.get_array('yc','S','sx')
+        self.area = sum(sx*yc*2.*S)
+        if self.area<1e-5: # check body area
+            raise ValueError("PanelArray must have positive area.\n"+
+                             "Check that your panels wrap clockwise.")
+
     ### Flow solver
 
     def solve_gamma(self,alpha=0,kutta=[]):
@@ -155,7 +166,7 @@ class PanelArray(object):
         Outputs:
         gamma of the PanelArray is updated.
 
-        Examples:
+        Example:
         foil = vp.make_jfoil(N=32)                    # make a Panel array
         foil.solve_gamma(alpha=0.1, kutta=[(0,-1)])      # solve for gamma
         foil.plot_flow()                                 # plot the flow
@@ -244,7 +255,7 @@ class PanelArray(object):
         Outputs:
         plot of the flow vectors, velocity magnitude contours, and the panels.
 
-        Examples:
+        Example:
         circle = vp.make_circle(N=32)      # make a Panel array
         circle.solve_gamma(alpha=0.1)      # solve for Panel strengths
         circle.plot_flow()                 # plot the flow
@@ -279,7 +290,7 @@ class PanelArray(object):
         Inputs:
         style -- a string defining the matplotlib style
 
[…]
