### 2019-03-15 8391c3e
Add sharp argument to vp.spline
