### 2019-03-15 2f2bc01
Update VortexPanel.py
