### 2019-03-16 41bb38c
Update VortexPanel.py
