### 2019-04-01 #45 @weymouth
I think this is a great repository and a model for other similar projects. The only things I notice are

1. The write-up doesn't make the development "story" clear.
1. The usage of the code in the repository is well explained. Would it be helpful to explain how this ties in with the other parts of the class? For example, is this supposed to be their first time seeing any of these concepts? What do they do with this experience afterward?
1. I don't see the version number. (Which I only mention because of the checklist.)
