### 2019-04-04 #45 @mesnardo
Hi @weymouth,

Thank you for your review of AeroPython! We'll work on your suggestions.

(@Juanlu001, looking forward to your review.)
