### 2019-04-16 #45 @kyleniemeyer
Hi @Juanlu001, just wanted to check in on the status of your review.

@mesnardo have you made changes according to @weymouth's suggestions?
