### 2019-04-20 #45 @labarba
Hi @Juanlu001 — we're eager to get  your review of the AeroPython module, to make a revision and get it to publication-ready status. Would you be able to get to this soon?
