### 2019-04-21 #45 @Juanlu001
It was quick because I already knew the course from the time it was released :)

Non-blocking comments:

* There is no version number, so I left that unchecked
* There is a statement of need in the paper but not in the repository itself - I understand that there are two checkboxes for this reason, so I left the first unchecked as well
* Perhaps this is a bit presumptuous on my side, but I would add a little footnote somewhere stating somethink like "not to be confused with another Python course for aeronautical engineers in Spanish language, with the same name but different content" (refs: https://github.com/AeroPython/Curso_AeroPython/, https://twitter.com/LorenaABarba/status/464041427169583104)
