### 2019-04-27 150d360
Add global version and others
Also:

* Remove version markdown in each notebook.
* Add global version for entire repository as a badge in the README (version set to `1.0`). (I will link the badge to the future GitHub release.)
* Update license info in header of each notebook (MIT -> BSD 3-Clause).
* Add license badges in README.
* Re-format README (size of section titles, one sentence per line, etc.).
