### 2019-04-29 7c38ebb
Update VortexPanel.py
Add condition for general split PanelArrays to fix offset body error message bug.
