### 2019-05-01 121a215
Add statement of need in readme
