### 2019-05-01 3d4b478
Mention other AeroPython repo in readme
