### 2019-05-01 c9fd058
Improve write-up of story
Add sentence on how AeroPython is used at GWU.
Add sentence on how AeroPython can be used for a full-semester course.
