### 2019-05-01 #45 @mesnardo
Thank you @weymouth and @Juanlu001 for your feedback!

Modifications based on your suggestions are available in the branch [`jose_revision`](https://github.com/barbagroup/AeroPython/tree/jose_revision) ; their is an open PR ([#44](https://github.com/barbagroup/AeroPython/pull/44)) to merge these commits in the `master` branch.

* We added a version number to the GitHub repository; the GitHub release will be done once the revision is finished.
* We added a note at the bottom of the README to mention the other [AeroPython](https://github.com/AeroPython/Curso_AeroPython/) repository.
* We modified the write-up to mention how AeroPython can be used as a full-semester course.
