### 2019-05-01 #45 @labarba
Regarding the Statement of Need, the JOSE editors decided it would only be required in the paper (not the documentation), but unfortunately the checklists haven't been updated yet!
