### 2019-05-13 #45 @kyleniemeyer
@labarba @mesnardo  a few minor comments on the paper itself:

- Could you add full affiliation details (city, state, country)?
- I believe the final bullet in the list of lessons should start with "Assignment 3"
- I just submitted a small fix to the paper near the end; multiple references in the same square brackets should be separated by semicolons: barbagroup/AeroPython#45
- the Catrambone reference is missing the DOI: 10.1037/0096-3445.127.4.355
- the Margulieux reference is missing the DOI: 10.1080/08993408.2016.1144429

Once you make those fixes, please archive the repo and provide the DOI. Almost done!
