### 2019-05-14 8209e1a
Fix last bullet in list of lessons
