### 2019-05-14 64b97fc
Add missing DOIs in bibliography
