+++ paper.bib
@@ -26,7 +26,8 @@
 	Publisher = {Taylor \& Francis},
 	Title = {Employing subgoals in computer programming education},
 	Volume = {26},
-	Year = {2016}}
+	Year = {2016},
+	Bdsk-Url-1 = {https://doi.org/10.1080/08993408.2016.1144429}}
 
 @article{catrambone1998,
 	Author = {Catrambone, Richard},
@@ -38,7 +39,8 @@
 	Publisher = {American Psychological Association},
 	Title = {The subgoal learning model: Creating better examples so that students can solve novel problems.},
 	Volume = {127},
-	Year = {1998}}
+	Year = {1998},
+	Bdsk-Url-1 = {https://doi.org/10.1037/0096-3445.127.4.355}}
 
 @article{chen2015worked,
 	Author = {Chen, Ouhao and Kalyuga, Slava and Sweller, John},
