### 2019-05-14 #45 @mesnardo
> @labarba @mesnardo a few minor comments on the paper itself:
>
> * Could you add full affiliation details (city, state, country)?

Done in [35cf51d](https://github.com/barbagroup/AeroPython/commit/35cf51df75dae985cfd2dae20bb1f68b71a0a51b).

> * I believe the final bullet in the list of lessons should start with "Assignment 3"

Done in [8209e1a](https://github.com/barbagroup/AeroPython/commit/8209e1a336f02e9ee7ce2821673515b803485b09).

> * I just submitted a small fix to the paper near the end; multiple references in the same square brackets should be separated by semicolons: [barbagroup/AeroPython#45](https://github.com/barbagroup/AeroPython/pull/45)

Thank you.

> * the Catrambone reference is missing the DOI: 10.1037/0096-3445.127.4.355
> * the Margulieux reference is missing the DOI: 10.1080/08993408.2016.1144429

Added mission DOIs in [64b97fc](https://github.com/barbagroup/AeroPython/commit/64b97fc62b202058854033d74e956b18e190e70e).

>
> Once you make those fixes, please archive the repo and provide the DOI. Almost done!
