### 2019-05-14 #45 @kyleniemeyer
@mesnardo looks like those DOIs were not actually added properly (used a bibdesk URL field rather than DOI field); I fixed in PR barbagroup/AeroPython#46

That should be the final edit necessary; please merge and then deposit for a DOI.
