### 2019-05-14 201dfbe
replace singular "module" by "set of lessons"
we present the collection as a set of 4 modules
