### 2019-05-15 #45 @kyleniemeyer
@labarba @mesnardo I'm not sure why the reference checking isn't working, but this looks good to go to me. Could you archive and provide me with the DOI?
