### 2019-05-17 #45 @labarba
👋 @weymouth, @Juanlu001 — We have now released v1.0 and updated the version number here, so you can tick off that item in your checklist. Many thanks!
