### 2019-05-17 #45 @labarba
On JOSS, yes. But here, we never set up an EiC team.
