### 2019-05-17 #45 @labarba
I was just fiddling to create a new team (which I did), but can't find how to give it Admin powers. Gah.
