### 2019-05-17 #45 @kyleniemeyer
Congrats to @labarba and @mesnardo on your submissions's publication in JOSE! Thanks to @weymouth and @Juanlu001 for reviewing!
